backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
langchain
langchain-community
langchain-openai
langchain-neo4j
neo4j
chromadb
python-pptx
pydantic
tqdm
requests